Fixed andReturn method of TerminalLongMediator.

+++ Source/ch/nolix/core/validator2/TerminalLongMediator.java
@@ -21,7 +21,7 @@ public final class TerminalLongMediator extends Mediator {
 	}
 	
 	//method
-	public double andReturn() {
+	public long andReturn() {
 		return argument;
 	}
 }
